PAAC-154: Split model classes out to own files. Moved implicit casts to package Utilities. Refactored Contribution
