+++ app/models/Contribution.scala
@@ -20,8 +20,7 @@ import play.api.libs.functional.syntax._
 import play.api.libs.json.Reads._
 import play.api.libs.json._
 
-sealed trait CalculationParam
-sealed trait PensionCalculatorValue {
+trait PensionCalculatorValue {
   def isEmpty(): Boolean
   def definedBenefit():  Option[Long]
   def moneyPurchase():  Option[Long]
@@ -29,217 +28,99 @@ sealed trait PensionCalculatorValue {
 }
 
 /**
-  Calculator inputs for a single year
- */
-case class InputAmounts(definedBenefit: Option[Long] = None, 
-                        moneyPurchase: Option[Long] = None, 
-                        income: Option[Long] = None,
-                        triggered: Option[Boolean] = None) extends PensionCalculatorValue {
-  def isEmpty() : Boolean = {
-    !definedBenefit.isDefined && !moneyPurchase.isDefined && !income.isDefined
-  }
-}
-
-/**
-  Simple date object representing a point in time within a pension period.
-*/
-case class PensionPeriod(year: Int, month: Int, day: Int) {
-  def <(that: PensionPeriod): Boolean = if (year == that.year && month == that.month && day == that.day) false else 
-                                    year < that.year || (year == that.year && month < that.month) || (year == that.year && month == that.month && day < that.day) 
-  def >(that: PensionPeriod): Boolean = if (year == that.year && month == that.month && day == that.day) false else 
-                                    year > that.year || (year == that.year && month > that.month) || (year == that.year && month == that.month && day > that.day) 
-  def <=(that: PensionPeriod): Boolean = if (year == that.year && month == that.month && day == that.day) true else this < that
-  def >=(that: PensionPeriod): Boolean = if (year == that.year && month == that.month && day == that.day) true else this > that
-  def isPeriod(s:PensionPeriod, e:PensionPeriod):Boolean = {
-    (this >= s) && (this <= e)
-  }
-
-  def isPeriod1(): Boolean = {
-    isPeriod(PensionPeriod.PERIOD_1_2015_START, PensionPeriod.PERIOD_1_2015_END)
-  }
-
-  def isPeriod2(): Boolean = {
-    isPeriod(PensionPeriod.PERIOD_2_2015_START, PensionPeriod.PERIOD_2_2015_END)
-  }
-
-  def taxYear(): Int = {
-    if (this < PensionPeriod(year, 4, 6) && this > PensionPeriod(year-1, 4, 6)) year-1
-    else if (this > PensionPeriod(year, 4, 6) && this < PensionPeriod(year+1, 4, 6)) year
-    else year
-  }
-}
-
-/**
   Contribution representing pension savings for a single pension period.
 */
-case class Contribution(taxPeriodStart: PensionPeriod, taxPeriodEnd: PensionPeriod, amounts: Option[InputAmounts]) extends CalculationParam {
+trait CalculationParam {
+  def taxPeriodStart(): PensionPeriod
+  def taxPeriodEnd(): PensionPeriod
+  def amounts(): Option[InputAmounts]
+
   /**
     Helper function for front-end. Should be converted to use messages.
   */
-  def taxYearLabel() : String = {
+  def taxYearLabel() : String =
     if (taxPeriodStart.year == 2015 && taxPeriodStart.month == 7 ||
-        taxPeriodEnd.year == 2016 && taxPeriodEnd.month == 4) {
+        taxPeriodEnd.year == 2016 && taxPeriodEnd.month == 4)
       s"2015/16 P2"  
-    } else if (taxPeriodStart.year == 2015 && taxPeriodStart.month == 4 ||
-               taxPeriodEnd.year == 2015 && taxPeriodEnd.month == 7) {
+    else if (taxPeriodStart.year == 2015 && taxPeriodStart.month == 4 ||
+             taxPeriodEnd.year == 2015 && taxPeriodEnd.month == 7)
       s"2015/16 P1"  
-    } else {
+    else
       s"${taxPeriodStart.year}/${taxPeriodEnd.year.toString().drop(2)}"
-    }
-  }
 
   /**
     Helper function for tests.
   */
   def label() : String = {
     val beforeAfter = if (amounts.getOrElse(InputAmounts()).triggered.getOrElse(false)) "A" else "B"
-    if (isPeriod2) {
+    if (isPeriod2)
       s"15/16 P2 $beforeAfter"  
-    } else if (isPeriod1) {
+    else if (isPeriod1)
       s"15/16 P1 $beforeAfter"  
-    } else {
+    else
       s"${taxPeriodStart.year.toString().drop(2)}/${taxPeriodEnd.year.toString().drop(2)}   "
-    }
   }
   
-  def isEmpty(): Boolean = {
+  def isEmpty(): Boolean =
     !amounts.isDefined || (amounts.isDefined && amounts.get.isEmpty)
-  }
 
-  def isPeriod1(): Boolean = {
+  def isPeriod1(): Boolean =
     taxPeriodStart.isPeriod1 && taxPeriodEnd.isPeriod1
-  }
 
-  def isPeriod2(): Boolean = {
+  def isPeriod2(): Boolean =
     taxPeriodStart.isPeriod2 && taxPeriodEnd.isPeriod2
-  }
 
-  def + (that:Contribution): Contribution = {
-    if (amounts.isDefined && that.amounts.isDefined) {
-      val thisAmounts = amounts.get
-      val thatAmounts = that.amounts.get
-      val db = thisAmounts.definedBenefit.map((v:Long)=>v+thatAmounts.definedBenefit.getOrElse(0L)).getOrElse(thatAmounts.definedBenefit.getOrElse(0L))
-      val dc = thisAmounts.moneyPurchase.map((v:Long)=>v+thatAmounts.moneyPurchase.getOrElse(0L)).getOrElse(thatAmounts.moneyPurchase.getOrElse(0L))
-      this.copy(amounts=Some(InputAmounts(db,dc)))
-    } else {
-      this
-    }
-  }
-
-  def isGroup1(): Boolean = {
+  def isGroup1(): Boolean =
     amounts.isDefined && 
     (isPeriod1() || isPeriod2()) &&
-    ( !amounts.get.moneyPurchase.isDefined &&
-     ( !amounts.get.triggered.isDefined|| !amounts.get.triggered.get))
-  }
+    (!amounts.get.moneyPurchase.isDefined && !isTriggered)
 
-  def isGroup2(): Boolean = {
+  def isGroup2(): Boolean =
     amounts.isDefined && 
     (isPeriod1() || isPeriod2()) && 
     amounts.get.moneyPurchase.isDefined
-  }
 
-  def isGroup3(): Boolean = {
+  def isGroup3(): Boolean =
     amounts.isDefined && 
     (isPeriod2() && 
-    amounts.get.triggered.getOrElse(false) && 
+    isTriggered && 
     amounts.get.moneyPurchase.isDefined &&
     amounts.get.definedBenefit.isDefined)
-  }
 
-  def isTriggered(): Boolean = {
+  def isTriggered(): Boolean =
     (for {
       inputs <- amounts
       triggered <- inputs.triggered
     } yield triggered) getOrElse false
-  }
 
-  def definedBenefit(): Long = {
+  def definedBenefit(): Long =
     (for {
       inputs <- amounts
       definedBenefit <- inputs.definedBenefit
     } yield definedBenefit) getOrElse 0L
-  }
   
-  def moneyPurchase(): Long = {
+  def moneyPurchase(): Long =
     (for {
       inputs <- amounts
       moneyPurchase <- inputs.moneyPurchase
     } yield moneyPurchase) getOrElse 0L
-  }
+
+  def income(): Long =
+    (for {
+      inputs <- amounts
+      income <- inputs.income
+    } yield income) getOrElse 0L
 }
 
-object PensionPeriod {
-  // Unlike front-end backend must have fixed supported start and end years
-  // as calculation rules are very dependant on a varying set of rules for each year
-  val EARLIEST_YEAR_SUPPORTED:Int = 2008
-  val LATEST_YEAR_SUPPORTED:Int = 2100
-
-  // Constants for pension period json validation
-  val MIN_MONTH_VALUE:Int = 1
-  val MIN_DAY_VALUE:Int = 1
-  val MAX_MONTH_VALUE:Int = 12
-  val MAX_DAY_VALUE:Int = 31
-
-  // Helpers for constructing pension contributions
-  val PERIOD_1_2015_START = PensionPeriod(2015, 4, 6)
-  val PERIOD_1_2015_END = PensionPeriod(2015, 7, 8)
-  val PERIOD_2_2015_START = PensionPeriod(2015, 7, 9)
-  val PERIOD_2_2015_END = PensionPeriod(2016, 4, 5)  
-
-  implicit val taxPeriodWrites: Writes[PensionPeriod] = (
-    (JsPath \ "year").write[Int] and
-    (JsPath \ "month").write[Int] and
-    (JsPath \ "day").write[Int]
-  )(unlift(PensionPeriod.unapply))
-
-  implicit val taxPeriodReads: Reads[PensionPeriod] = (
-    (JsPath \ "year").read[Int](min(EARLIEST_YEAR_SUPPORTED)) and
-    (JsPath \ "month").read[Int](min(MIN_MONTH_VALUE) keepAnd max(MAX_MONTH_VALUE)) and
-    (JsPath \ "day").read[Int](min(MIN_DAY_VALUE) keepAnd max(MAX_DAY_VALUE))
-  )(PensionPeriod.apply _)
-}
-
-object InputAmounts {
-  implicit val inputAmountsWrites: Writes[InputAmounts] = (
-    (JsPath \ "definedBenefit").write[Option[Long]] and
-    (JsPath \ "moneyPurchase").write[Option[Long]] and
-    (JsPath \ "income").write[Option[Long]] and
-    (JsPath \ "triggered").write[Option[Boolean]]
-  )(unlift(InputAmounts.unapply))
-
-  implicit val inputAmountsReads: Reads[InputAmounts] = (
-    (JsPath \ "definedBenefit").readNullable[Long](min(0L)) and
-    (JsPath \ "moneyPurchase").readNullable[Long](min(0L)) and
-    (JsPath \ "income").readNullable[Long](min(0L)) and
-    (JsPath \ "triggered").readNullable[Boolean]
-  )(InputAmounts.apply(_: Option[Long], _: Option[Long], _: Option[Long], _: Option[Boolean]))
-
-  /**
-    Simplified apply function
-  */
-  def apply(definedBenefit: Long, moneyPurchase: Long, income: Long) : InputAmounts = {
-    InputAmounts(Some(definedBenefit), Some(moneyPurchase), Some(income), None)
-  }
-
-  /**
-    Simplified apply function
-  */
-  def apply(definedBenefit: Long, moneyPurchase: Long) : InputAmounts = {
-    InputAmounts(Some(definedBenefit), Some(moneyPurchase), None, None)
-  }
-
-  /**
-    Simplified apply function
-  */
-  def apply(definedBenefit: Long) : InputAmounts = {
-    InputAmounts(Some(definedBenefit), None, None, None)
-  }
+case class Contribution(taxPeriodStart: PensionPeriod, taxPeriodEnd: PensionPeriod, amounts: Option[InputAmounts]) extends CalculationParam {
+  def + (that:Contribution): Contribution =
+    if (amounts.isDefined && that.amounts.isDefined)
+      this.copy(amounts=Some(InputAmounts((definedBenefit + that.definedBenefit),(moneyPurchase + that.moneyPurchase),(income + that.income))))
+    else
+      this
 }
 
 object Contribution {
-  import calculators.internal.Utilities._
-
   implicit val contributionWrites: Writes[Contribution] = (
     (JsPath \ "taxPeriodStart").write[PensionPeriod] and
     (JsPath \ "taxPeriodEnd").write[PensionPeriod] and
@@ -251,18 +132,6 @@ object Contribution {
     (JsPath \ "taxPeriodEnd").read[PensionPeriod] and
     (JsPath \ "amounts").readNullable[InputAmounts]
   )(Contribution.apply(_:PensionPeriod, _:PensionPeriod, _:Option[InputAmounts]))
-
-  /**
-    Implicit cast function from Contribution to SummaryResultsTuple.
-    Used when calculating actual unused allowance.
-  */
-  implicit def convert(c: Contribution)(implicit calculator:calculators.internal.SummaryCalculator): SummaryResultsTuple = {
-    implicit val contribution = c
-    contribution match {
-      case _ if c.isPeriod1 || c.isPeriod2 => (2015, calculator.exceedingAllowance, calculator.unusedAllowance)
-      case _ => (contribution.taxPeriodStart.year, calculator.exceedingAllowance, calculator.unusedAllowance)
-    }
-  }
 
   /**
     Simplified apply function
